+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Changed
 ### Fixed
 
+## [0.1.8] - 2019-08-31
+### Added
+- A new method `case_string_values` on Column to get a list of string values from a `CASE` expression.
+
 ## [0.1.7] - 2019-06-19
 ### Added
 - `ordered_query_parameter_definitions` method for Reports. Returns parameter definitions ordered first by UI priority, then by caption.